# kv: include conflicting request information in latch manager traces/logs