… pkg/kv/kvserver/concurrency/lock_table_test.go
… pkg/kv/kvserver/concurrency/lock_table_waiter_test.go
… pkg/kv/kvserver/spanlatch/manager.go
… pkg/kv/kvserver/spanlatch/manager_test.go