@lyang24 2023-12-12
Hi @nvanbenschoten , I rebased this pr to the recent master that reflects the refactor of request safeformat. Do you mind take a look at this pr when you have time?