@lyang24 2024-01-03
seems like this had a small regression on benchmarks :(
```
./benchstat ~/Desktop/master.txt ~/Desktop/mybrach.txt 
goos: darwin
goarch: arm64
                                              │ /Users/lanqing/Desktop/master.txt │ /Users/lanqing/Desktop/mybrach.txt  │
                                              │              sec/op               │   sec/op     vs base                │
LatchManagerReadOnlyMix/size=1-12                                     130.5n ± 4%   144.9n ± 1%  +11.03% (p=0.000 n=10)
LatchManagerReadOnlyMix/size=4-12                                     131.5n ± 2%   145.5n ± 1%  +10.68% (p=0.000 n=10)
LatchManagerReadOnlyMix/size=16-12                                    133.3n ± 3%   143.9n ± 1%   +7.95% (p=0.000 n=10)
LatchManagerReadOnlyMix/size=64-12                                    135.0n ± 2%   145.7n ± 2%   +7.97% (p=0.000 n=10)
LatchManagerReadOnlyMix/size=128-12                                   136.4n ± 1%   144.4n ± 2%   +5.83% (p=0.000 n=10)
LatchManagerReadOnlyMix/size=256-12                                   137.9n ± 1%   146.9n ± 1%   +6.53% (p=0.000 n=10)
LatchManagerReadWriteMix/readsPerWrite=0-12                           203.5n ± 0%   214.9n ± 3%   +5.63% (p=0.000 n=10)
LatchManagerReadWriteMix/readsPerWrite=1-12                           186.4n ± 2%   198.2n ± 2%   +6.30% (p=0.000 n=10)
LatchManagerReadWriteMix/readsPerWrite=4-12                           186.4n ± 2%   197.1n ± 1%   +5.74% (p=0.000 n=10)
LatchManagerReadWriteMix/readsPerWrite=16-12                          179.0n ± 4%   188.7n ± 1%   +5.39% (p=0.001 n=10)
LatchManagerReadWriteMix/readsPerWrite=64-12                          120.7n ± 4%   130.6n ± 1%   +8.21% (p=0.000 n=10)
LatchManagerReadWriteMix/readsPerWrite=128-12                         110.6n ± 4%   120.1n ± 2%   +8.50% (p=0.000 n=10)
LatchManagerReadWriteMix/readsPerWrite=256-12                         104.0n ± 5%   115.2n ± 2%  +10.77% (p=0.000 n=10)
geomean                                                               142.7n        153.7n        +7.71%

                                              │ /Users/lanqing/Desktop/master.txt │ /Users/lanqing/Desktop/mybrach.txt │
                                              │               B/op                │    B/op      vs base               │
LatchManagerReadOnlyMix/size=1-12                                      192.0 ± 0%    208.0 ± 0%  +8.33% (p=0.000 n=10)
LatchManagerReadOnlyMix/size=4-12                                      192.0 ± 0%    208.0 ± 0%  +8.33% (p=0.000 n=10)
LatchManagerReadOnlyMix/size=16-12                                     192.0 ± 0%    208.0 ± 0%  +8.33% (p=0.000 n=10)
LatchManagerReadOnlyMix/size=64-12                                     192.0 ± 0%    208.0 ± 0%  +8.33% (p=0.000 n=10)
LatchManagerReadOnlyMix/size=128-12                                    192.0 ± 0%    208.0 ± 0%  +8.33% (p=0.000 n=10)
LatchManagerReadOnlyMix/size=256-12                                    192.0 ± 0%    208.0 ± 0%  +8.33% (p=0.000 n=10)
LatchManagerReadWriteMix/readsPerWrite=0-12                            192.0 ± 0%    208.0 ± 0%  +8.33% (p=0.000 n=10)
LatchManagerReadWriteMix/readsPerWrite=1-12                            192.0 ± 0%    208.0 ± 0%  +8.33% (p=0.000 n=10)
LatchManagerReadWriteMix/readsPerWrite=4-12                            192.0 ± 0%    208.0 ± 0%  +8.33% (p=0.000 n=10)
LatchManagerReadWriteMix/readsPerWrite=16-12                           192.0 ± 0%    208.0 ± 0%  +8.33% (p=0.000 n=10)
LatchManagerReadWriteMix/readsPerWrite=64-12                           192.0 ± 0%    208.0 ± 0%  +8.33% (p=0.000 n=10)
LatchManagerReadWriteMix/readsPerWrite=128-12                          192.0 ± 0%    208.0 ± 0%  +8.33% (p=0.000 n=10)
LatchManagerReadWriteMix/readsPerWrite=256-12                          192.0 ± 0%    208.0 ± 0%  +8.33% (p=0.000 n=10)
geomean                                                                192.0         208.0       +8.33%

                                              │ /Users/lanqing/Desktop/master.txt │ /Users/lanqing/Desktop/mybrach.txt  │
                                              │             allocs/op             │ allocs/op   vs base                 │
LatchManagerReadOnlyMix/size=1-12                                      1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
LatchManagerReadOnlyMix/size=4-12                                      1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
LatchManagerReadOnlyMix/size=16-12                                     1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
LatchManagerReadOnlyMix/size=64-12                                     1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
LatchManagerReadOnlyMix/size=128-12                                    1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
LatchManagerReadOnlyMix/size=256-12                                    1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
LatchManagerReadWriteMix/readsPerWrite=0-12                            1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
LatchManagerReadWriteMix/readsPerWrite=1-12                            1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
LatchManagerReadWriteMix/readsPerWrite=4-12                            1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
LatchManagerReadWriteMix/readsPerWrite=16-12                           1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
LatchManagerReadWriteMix/readsPerWrite=64-12                           1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
LatchManagerReadWriteMix/readsPerWrite=128-12                          1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
LatchManagerReadWriteMix/readsPerWrite=256-12                          1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                1.000        1.000       +0.00%
```